Reuse `Card` component for Readme section
